Move some code
